Relocate some readers methods
Some methods in Y2Users::Linux::BaseReaders make no sense there since
they are not compatible with Y2Users::Linux::LocalReader. Thus, they
have been placed in Y2User::Linux::Reader until they allow reading from
a specific location.
