+++ src/lib/y2users/linux/base_reader.rb
@@ -18,7 +18,6 @@
 # find current contact information at www.suse.com.
 
 require "yast"
-require "yast2/execute"
 require "abstract_method"
 require "y2users/config"
 require "y2users/login_config"
@@ -29,7 +28,6 @@ require "y2users/linux/useradd_config_reader"
 require "users/ssh_authorized_keyring"
 
 Yast.import "Autologin"
-Yast.import "MailAliases"
 
 module Y2Users
   module Linux
@@ -43,12 +41,7 @@ module Y2Users
       def read
         Config.new.tap do |config|
           read_elements(config)
-          read_root_aliases(config)
           read_passwords(config)
-          read_home_permissions(config)
-          read_authorized_keys(config)
-          read_useradd_config(config)
-          read_login(config)
         end
       end
 
@@ -89,18 +82,6 @@ module Y2Users
       #   @return [String] loaded groups from the system
       abstract_method :load_groups
 
-      # Set root aliases (i.e., users receiving system mails)
-      #
-      # @see Yast::MailAliases#GetRootAlias
-      #
-      # @param config [Config]
-      def read_root_aliases(config)
-        Yast::MailAliases.GetRootAlias.split(", ").each do |name|
-          user = config.users.by_name(name)
-          user.receive_system_mail = true if user
-        end
-      end
-
       # Parses the content retrieved by {#load_passwords} and sets user passwords
       #
       # @see Parsers::Shadow#parse
@@ -123,58 +104,6 @@ module Y2Users
       # @!method load_passwords
       #   @return [String] loaded passwords from the system
       abstract_method :load_passwords
-
-      # Command for reading home directory permissions
-      STAT = "/usr/bin/stat".freeze
-      private_constant :STAT
-
-      # Reads home permissions
-      #
-      # @return [Array<Y2Users::User>]
-      def read_home_permissions(config)
-        config.users.reject(&:system?).each do |user|
-          next unless user.home && Dir.exist?(user.home.path)
-
-          user.home.permissions = Yast::Execute.locally!(
-            STAT, "--printf", "%a", user.home.path,
-            stdout: :capture
-          )
-        end
-      end
-
-      # Reads users authorized keys
-      #
-      # @see Yast::Users::SSHAuthorizedKeyring#read_keys
-      # @return [Array<Y2Users::User>]
-      def read_authorized_keys(config)
-        config.users.each do |user|
-          next unless user.home
-
-          user.authorized_keys = Yast::Users::SSHAuthorizedKeyring.new(user.home.path).read_keys
-        end
-      end
-
-      # Reads the configuration for useradd
-      #
-      # @param config [Config]
-      def read_useradd_config(config)
-        config.useradd = UseraddConfigReader.new.read
-      end
-
-      # Reads the login information
-      #
-      # @param config [Config]
-      def read_login(config)
-        Yast::Autologin.Read
-
-        return unless Yast::Autologin.used
-
-        login = LoginConfig.new
-        login.autologin_user = config.users.by_name(Yast::Autologin.user)
-        login.passwordless = Yast::Autologin.pw_less
-
-        config.login = login
-      end
     end
   end
 end
