+++ src/lib/y2users/linux/reader.rb
@@ -20,11 +20,93 @@
 require "yast2/execute"
 require "y2users/linux/base_reader"
 
+Yast.import "MailAliases"
+
 module Y2Users
   module Linux
     # Reads users configuration from the system using `getent` command.
     class Reader < BaseReader
-    private # rubocop:disable Layout/IndentationWidth
+      # @see BaseReader#read
+      # @note some of these #read_* methods could be moved to the
+      #   {BaseReader#read} once they allow reading from a specific location
+      #   (i.e., compatible with {LocalReader} too)
+      def read
+        config = super
+
+        read_root_aliases(config)
+        read_home_permissions(config)
+        read_authorized_keys(config)
+        read_useradd_config(config)
+        read_login(config)
+
+        config
+      end
+
+    private
+
+      # Set root aliases (i.e., users receiving system mails)
+      #
+      # @see Yast::MailAliases#GetRootAlias
+      #
+      # @param config [Config]
+      def read_root_aliases(config)
+        Yast::MailAliases.GetRootAlias.split(", ").each do |name|
+          user = config.users.by_name(name)
+          user.receive_system_mail = true if user
+        end
+      end
+
+      # Command for reading home directory permissions
+      STAT = "/usr/bin/stat".freeze
+      private_constant :STAT
+
+      # Reads home permissions
+      #
+      # @return [Array<Y2Users::User>]
+      def read_home_permissions(config)
+        config.users.reject(&:system?).each do |user|
+          next unless user.home && Dir.exist?(user.home.path)
+
+          user.home.permissions = Yast::Execute.locally!(
+            STAT, "--printf", "%a", user.home.path,
+            stdout: :capture
+          )
+        end
+      end
+
+      # Reads users authorized keys
+      #
+      # @see Yast::Users::SSHAuthorizedKeyring#read_keys
+      # @return [Array<Y2Users::User>]
+      def read_authorized_keys(config)
+        config.users.each do |user|
+          next unless user.home
+
+          user.authorized_keys = Yast::Users::SSHAuthorizedKeyring.new(user.home.path).read_keys
+        end
+      end
+
+      # Reads the configuration for useradd
+      #
+      # @param config [Config]
+      def read_useradd_config(config)
+        config.useradd = UseraddConfigReader.new.read
+      end
+
+      # Reads the login information
+      #
+      # @param config [Config]
+      def read_login(config)
+        Yast::Autologin.Read
+
+        return unless Yast::Autologin.used
+
+        login = LoginConfig.new
+        login.autologin_user = config.users.by_name(Yast::Autologin.user)
+        login.passwordless = Yast::Autologin.pw_less
+
+        config.login = login
+      end
 
       # Loads entries from `passwd` database
       #
