+++ test/lib/y2users/linux/base_reader_test.rb
@@ -25,34 +25,20 @@ require "y2users/config"
 require "y2users/linux/base_reader"
 
 describe Y2Users::Linux::BaseReader do
-  around do |example|
-    # Let's use test/fixtures/home as src root for reading authorized keys from there
-    change_scr_root(FIXTURES_PATH.join("home")) { example.run }
-  end
-
   describe "#read" do
-    let(:passwd_content) { File.read(File.join(FIXTURES_PATH, "/root2/etc/passwd")) }
-    let(:group_content)  { File.read(File.join(FIXTURES_PATH, "/root2/etc/group")) }
-    let(:shadow_content) { File.read(File.join(FIXTURES_PATH, "/root2/etc/shadow")) }
+    let(:passwd_content) { File.read(File.join(FIXTURES_PATH, "/root/etc/passwd")) }
+    let(:group_content)  { File.read(File.join(FIXTURES_PATH, "/root/etc/group")) }
+    let(:shadow_content) { File.read(File.join(FIXTURES_PATH, "/root/etc/shadow")) }
     let(:root_home) { FIXTURES_PATH.join("home", "root").to_s }
     let(:expected_root_auth_keys) { authorized_keys_from(root_home) }
 
     before do
+      allow(subject.log).to receive(:warn)
+
       # mock Yast::Execute calls and provide file content from fixture
       allow(subject).to receive(:load_users).and_return(passwd_content)
       allow(subject).to receive(:load_groups).and_return(group_content)
       allow(subject).to receive(:load_passwords).and_return(shadow_content)
-
-      allow(subject.log).to receive(:warn)
-      allow(Yast::MailAliases).to receive(:GetRootAlias).and_return("games, unknown, news")
-
-      # mocks to check reading of home permissions
-      allow(Dir).to receive(:exist?)
-      allow(Dir).to receive(:exist?).with("/home/a_user").and_return(true)
-      allow(Yast::Execute).to receive(:locally!)
-      allow(Yast::Execute).to receive(:locally!)
-        .with("/usr/bin/stat", any_args, "/home/a_user", stdout: :capture)
-        .and_return("700")
     end
 
     it "generates a config with read data" do
@@ -60,8 +46,8 @@ describe Y2Users::Linux::BaseReader do
 
       expect(config).to be_a(Y2Users::Config)
 
-      expect(config.users.size).to eq 19
-      expect(config.groups.size).to eq 7
+      expect(config.users.size).to eq 18
+      expect(config.groups.size).to eq 37
 
       root_user = config.users.root
       expect(root_user.uid).to eq "0"
@@ -70,7 +56,6 @@ describe Y2Users::Linux::BaseReader do
       expect(root_user.primary_group.name).to eq "root"
       expect(root_user.password.value.encrypted?).to eq true
       expect(root_user.password.value.content).to match(/^\$6\$pL/)
-      expect(root_user.authorized_keys).to eq(expected_root_auth_keys)
     end
 
     it "logs warning if password found for not existing user" do
@@ -79,22 +64,6 @@ describe Y2Users::Linux::BaseReader do
       expect(subject.log).to receive(:warn).with(/Found password for.*fakeuser./)
 
       subject.read
-    end
-
-    it "sets root aliases" do
-      config = subject.read
-
-      root_aliases = config.users.select(&:receive_system_mail?)
-
-      expect(root_aliases.size).to eq 2
-      expect(root_aliases.map(&:name)).to contain_exactly("games", "news")
-    end
-
-    it "sets home permissions" do
-      config = subject.read
-
-      user = config.users.by_name("a_user")
-      expect(user.home.permissions).to eq("700")
     end
   end
 end
